requirements.txt
google-api-python-client
google-genai
python-dotenv
requests
pandas
pydantic
streamlit
rich
